migrating to nextjs from react
