+++ sdk-demo/package.json
@@ -15,6 +15,7 @@
     "@types/react": "^18.2.12",
     "@types/react-dom": "^18.2.5",
     "ethers": "^5.6.1",
+    "next": "^13.4.19",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
@@ -24,10 +25,8 @@
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "dev": "next dev",
+    "build": "next build"
   },
   "eslintConfig": {
     "extends": [
